revert package relocation
